Add validate_bundle method (#59)
* Add validate_bundle method

Add a method validate_bundle, that takes the validation setps and groups them together. By that I can reuse this code outside of the validator in qontract-reconcile. The goal is to be able to validate generated datafiles from templates withouth the need to actually generate the templates. This will be achieved by using the validator integration, with the templateTests.

* Fix format